Removing Export Screen & Updating QuickTrimProcessState
